+++ SectionKit/Tests/CollectionViewAdapter/CollectionViewContext/MainCollectionViewContextTests.swift
@@ -453,66 +453,3 @@ internal final class MainCollectionViewContextTests: XCTestCase {
         waitForExpectations(timeout: 1)
     }
 }
-
-private class MockCollectionView: UICollectionView {
-    // MARK: - reloadData
-
-    typealias ReloadDataBlock = () -> Void
-
-    lazy var _reloadData: ReloadDataBlock = { }
-
-    override func reloadData() {
-        _reloadData()
-    }
-
-    // MARK: - registerCell
-
-    typealias RegisterCellBlock = (AnyClass?, String) -> Void
-
-    lazy var _registerCell: RegisterCellBlock = { _, _ in }
-
-    override func register(_ cellClass: AnyClass?, forCellWithReuseIdentifier identifier: String) {
-        _registerCell(cellClass, identifier)
-    }
-
-    // MARK: - dequeueReusableCell
-
-    typealias DequeueReusableCellBlock = (String, IndexPath) -> UICollectionViewCell
-
-    lazy var _dequeueReusableCell: DequeueReusableCellBlock = { _, _ in UICollectionViewCell() }
-
-    override func dequeueReusableCell(
-        withReuseIdentifier identifier: String,
-        for indexPath: IndexPath
-    ) -> UICollectionViewCell {
-        _dequeueReusableCell(identifier, indexPath)
-    }
-
-    // MARK: - registerSupplementaryView
-
-    typealias RegisterSupplementaryViewBlock = (AnyClass?, String, String) -> Void
-
-    lazy var _registerSupplementaryView: RegisterSupplementaryViewBlock = { _, _, _ in }
-
-    override func register(
-        _ viewClass: AnyClass?,
-        forSupplementaryViewOfKind elementKind: String,
-        withReuseIdentifier identifier: String
-    ) {
-        _registerSupplementaryView(viewClass, elementKind, identifier)
-    }
-
-    // MARK: - dequeueReusableSupplementaryView
-
-    typealias DequeueReusableSupplementaryViewBlock = (String, String, IndexPath) -> UICollectionReusableView
-
-    lazy var _dequeueReusableSupplementaryView: DequeueReusableSupplementaryViewBlock = { _, _, _ in UICollectionReusableView() }
-
-    override func dequeueReusableSupplementaryView(
-        ofKind elementKind: String,
-        withReuseIdentifier identifier: String,
-        for indexPath: IndexPath
-    ) -> UICollectionReusableView {
-        _dequeueReusableSupplementaryView(elementKind, identifier, indexPath)
-    }
-}
